Update package dependencies and add new command implementations for URL and file handling
