+++ leojs.leo
@@ -603,6 +603,7 @@
 <v t="felix.20211018202009.1"><vh>@file src/core/decorators.ts</vh></v>
 <v t="felix.20230509194418.1"><vh>@file src/core/idle_time.ts</vh></v>
 <v t="felix.20210102012334.1"><vh>@file src/core/leoApp.ts</vh></v>
+<v t="felix.20250308203620.1"><vh>@file src/core/leoAppCommands.ts</vh></v>
 <v t="felix.20211211234842.1"><vh>@file src/core/leoAtFile.ts</vh></v>
 <v t="felix.20230802145823.1"><vh>@file src/core/leoCache.ts</vh></v>
 <v t="felix.20220429005433.1"><vh>@file src/core/leoChapters.ts</vh></v>
@@ -6780,7 +6781,7 @@ public async insertNode(p_node: Position | undefined, p_fromOutline: boolean, p_
 "ksuid": "^3.0.0",
 "lodash": "^4.17.21",
 "md5": "^2.3.0",
-"open": "^10.1.0",
+"open": "^8.4.2",
 "os-browserify": "^0.3.0",
 "pako": "^2.1.0",
 "path-browserify": "^1.0.1",
