+++ package.json
@@ -4649,7 +4649,7 @@
     "ksuid": "^3.0.0",
     "lodash": "^4.17.21",
     "md5": "^2.3.0",
-    "open": "^10.1.0",
+    "open": "^8.4.2",
     "os-browserify": "^0.3.0",
     "pako": "^2.1.0",
     "path-browserify": "^1.0.1",
